pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envleaks"
version = "0.1.0"
description = "Scan codebases, git history, and Docker images for accidentally exposed secrets"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "ExploitCraft", email = "exploitcraft@users.noreply.github.com" }]
keywords = ["security", "secrets", "leak", "scanner", "devops", "devsecops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "gitpython>=3.1",
]

[project.scripts]
envleaks = "envleaks.cli:main"

[project.urls]
Homepage = "https://github.com/ExploitCraft/envleaks"
Repository = "https://github.com/ExploitCraft/envleaks"
"Bug Tracker" = "https://github.com/ExploitCraft/envleaks/issues"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
